Бузникова Мария 3821ПИмэ

+++ src/fun.cpp
@@ -8,21 +8,15 @@ unsigned int faStr1(const char *str)
 	int count = 0;
 	int i = 0;
 	bool digit = false;
-	while (str[i]) 
-	{
-		if (isdigit(str[i]) && digit == false)
-		{
+	while (str[i]) {
+		if (isdigit(str[i]) && digit == false){
 			digit = true;
 		}
-		if (str[i] != ' ' && inWord == false)
-		{
+		if (str[i] != ' ' && inWord == false){
 			inWord = true;
 			count++;
-		}
-		else if (str[i] == ' ' && inWord == true)
-		{
-			if (digit == true)
-			{
+		} else if (str[i] == ' ' && inWord == true){
+			if (digit == true){
 				count--;
 				digit = false;
 			}
@@ -40,25 +34,17 @@ unsigned int faStr2(const char* str) {
 	int count = 0;
 	int i = 0;
 	bool Up = false;
-	while (str[i]) 
-	{
-		if (str[i] != ' ' && inWord == false && isupper(str[i]))
-		{
+	while (str[i]) {
+		if (str[i] != ' ' && inWord == false && isupper(str[i])){
 			inWord = true;
 			count++;
-		}
-		
-		else if (str[i] != ' ' && inWord == true && !islower(str[i]) && Up == false)
-		{
+		} else if (str[i] != ' ' && inWord == true && !islower(str[i]) && Up == false){
 			Up = true;
-		}
-		else if (str[i] == ' ' && inWord == true)
-		{
+		} else if (str[i] == ' ' && inWord == true){
 			inWord = false;
 			Up = false;
 		}
-		 if(Up==true)
-		 {
+		if(Up==true){
 			 Up = false;
 		 }
 		i++;
@@ -69,26 +55,22 @@ unsigned int faStr2(const char* str) {
 
 unsigned int faStr3(const char* str) {
 	bool inWord = false;
-	int count = 0;
-	int str_len = 0;
+	double count = 0;
+	double str_len = 0;
 	//int len = 0;
 	int i = 0;
-	while (str[i])
-	{
-		if (str[i] != ' ' && inWord == false)
-		{
+	while (str[i]){
+		if (str[i] != ' ' && inWord == false){
 			inWord = true;
 			count++;
 		}
-		if (str[i] != ' ' && inWord == true)
-		{
+		if (str[i] != ' ' && inWord == true){
 			str_len++;
-		}
-		else if (str[i] == ' ' && inWord == true)
+		} else if (str[i] == ' ' && inWord == true)
 			inWord = false;
 		i++;
 	}
-	double average = round((double)str_len / (double)count);
+	double average = round(str_len / count);
 	std::cout << average << std::endl;
     return 0;
 }
